refactor: extract TimelineControls and TimelineLegend components
Move navigation controls (~110 lines) and legend toggles (~130 lines)
from the Timeline component JSX into dedicated components.

- TimelineControls: nav buttons, date label, refresh, orientation
  toggle, fullscreen
- TimelineLegend: collapsible category toggles with colored dots,
  orientation-aware metrics sub-toggles

index.tsx: 1318 → 1081 lines

Total rewrite progress: 2770 → 1081 lines (61% reduction)
